# src/main/resources/api/data-repository-openapi.yaml
openapi: 3.0.1
info:
  title: Data Repository API
  description: >
    <details><summary>This document defines the REST API for the Terra Data Repository.</summary>
    <p>
    **Status: design in progress**

    There are a few top-level endpoints (besides some used by swagger):
     * / - generated by swagger: swagger API page that provides this
    documentation and a live UI for submitting REST requests
     * /status - provides the operational status of the service
     * /configuration - provides the basic configuration and information about the service
     * /api - is the authenticated and authorized Data Repository API
     * /ga4gh/drs/v1 - is a transcription of the Data Repository Service API

    The API endpoints are organized by interface. Each interface is separately versioned.
    <p>
    **Notes on Naming**
    <p>
    All of the reference items are suffixed with \"Model\".
    Those names are used as the class names in the generated Java code.
    It is helpful to distinguish these model classes from other related classes, like the DAO classes and the operation classes.
    </details>
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://app.terra.bio/#terms-of-service
  version: 0.1.0
servers:
  - url: /
security:
  - googleoauth:
      - openid
      - email
      - profile
  - oidc:
      - openid
      - email
      - profile
paths:
...
  /api/resources/v1/profiles:
    get:
      tags:
        - profiles
        - resources
      description: >
        Returns a list of all of the billing profiles
      operationId: enumerateProfiles
      parameters:
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
            set.
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: The numbers of items to return.
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: List of billing profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnumerateBillingProfileModel'
    post:
      tags:
        - profiles
        - resources
      description: >
        Creates a new profile associated with a billing account. This is asynchronous.
        The final return is of type definitions/BillingProfileModel.
      operationId: createProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingProfileRequestModel'
        required: false
      responses:
        200:
          description: Redirect for profile successful creation
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status of profile creation job & url for polling in the
            response header
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - see error message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to create profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      x-codegen-request-body-name: billingProfileRequest
    put:
      tags:
        - profiles
        - resources
      description: >
        Update a billing profile by id. This is asynchronous.
        The final return is of type definitions/BillingProfileModel.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingProfileUpdateModel'
        required: true
      responses:
        200:
          description: Redirect for profile update result
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status of profile update job & url for polling in the response
            header
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to update profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - profile does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/resources/v1/profiles/{id}:
    get:
      tags:
        - profiles
        - resources
      description: Retrieve a billing profile by id
      operationId: retrieveProfile
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Billing profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingProfileModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see billing profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - billing profile id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
    delete:
      tags:
        - profiles
        - resources
      description: >
        Delete a billing profile by id. This is asynchronous.
        The final return is of type definitions/DeleteResponseModel.
      operationId: deleteProfile
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: deleteCloudResources
          in: query
          description: ADMIN ONLY - Delete all Azure cloud resources along with billing profile.
          schema:
            type: boolean
            default: false
          required: false
      responses:
        200:
          description: Redirect for profile delete result
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status of profile delete job & url for polling in the response
            header
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to delete profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - profile does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        409:
          description: Profile in use by at least one dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/resources/v1/profiles/{id}/policies:
    get:
      tags:
        - profiles
        - resources
      description: Retrieve the read and discover policies for the profile
      operationId: retrieveProfilePolicies
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - snapshot id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/resources/v1/profiles/{id}/policies/{policyName}/members:
    post:
      tags:
        - profiles
        - resources
      description: >
        Adds a member to the specified policy for the profile.

        NOTE: Change may take up to 60 seconds to take effect.
      operationId: addProfilePolicyMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyMemberRequest'
        required: true
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/ProfilePolicyName'
      responses:
        201:
          description: Policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        400:
          description: Bad request - invalid id, badly formed, or email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to update policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - profile id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/resources/v1/profiles/{id}/policies/{policyName}/members/{memberEmail}:
    delete:
      tags:
        - profiles
        - resources
      description: >
        Removes a member from the specified policy for the profile.

        NOTE: Change may take up to 60 seconds to take effect.
      operationId: deleteProfilePolicyMember
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: policyName
          in: path
          description: The relevant policy
          required: true
          schema:
            type: string
            enum:
              - owner
              - user
        - name: memberEmail
          in: path
          description: The email of the user to remove
          required: true
          schema:
            type: string
      responses:
        200:
          description: PolicyResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to update policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - snapshot id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
  /api/repository/v1/jobs/{id}:
    get:
      tags:
        - jobs
        - repository
      description: Retrieve a job's status by id
      operationId: retrieveJob
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      responses:
        200:
          description: Redirect for object successful creation
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - job id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
components:
  schemas:
...
    UniqueIdProperty:
      type: string
      format: uuid
      description: >
        Unique identifier for a dataset, snapshot, etc.
    ShortIdProperty:
      pattern: '[A-Za-z0-9_\\-]{22}'
      type: string
      description: >
        Unique identifier for flights, jobs, etc.
    BillingProfileRequestModel:
      required:
        - biller
        - profileName
      type: object
      properties:
        billingAccountId:
          type: string
          description: unique identifier of the billing account from Google
        profileName:
          type: string
          description: display name to be used for the profile;
        biller:
          type: string
          description: either “direct” or the reseller of the sub-account that sends
            the bill (like “on_spend”)
        description:
          type: string
          description: Free text that describes the profile
        cloudPlatform:
          $ref: '#/components/schemas/CloudPlatform'
        tenantId:
          $ref: '#/components/schemas/UniqueIdProperty'
        subscriptionId:
          $ref: '#/components/schemas/UniqueIdProperty'
        resourceGroupName:
          type: string
          description: an optional resource group name for Azure
        applicationDeploymentName:
          type: string
          description: an optional name for an application deployment for Azure
        id:
          deprecated: true
          type: string
          format: uuid
          description: >
            Deprecated: allowed for backwards compatibility.
            Recommendation: leave this unset for TDR to generate a universally unique identifier
            for your new billing profile.
    BillingProfileUpdateModel:
      required:
        - id
        - billingAccountId
        - description
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        billingAccountId:
          type: string
          description: Unique identifier of the billing account from Google
        description:
          type: string
          description: Free text that describes the profile
    BillingProfileModel:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        billingAccountId:
          type: string
          description: unique identifier of the billing account from Google
        profileName:
          type: string
          description: display name to be used for the profile;
        biller:
          type: string
          description: either “direct” or the reseller of the sub-account that sends
            the bill (like “on_spend”)
        description:
          type: string
          description: Free text that describes the profile
        cloudPlatform:
          $ref: '#/components/schemas/CloudPlatform'
        tenantId:
          $ref: '#/components/schemas/UniqueIdProperty'
        subscriptionId:
          $ref: '#/components/schemas/UniqueIdProperty'
        resourceGroupName:
          type: string
          description: an optional resource group name for Azure
        applicationDeploymentName:
          type: string
          description: an optional name for an application deployment for Azure
        createdDate:
          type: string
          description: Date the profile was created
        createdBy:
          type: string
          description: Identity of the user who originally created the profile
    EnumerateBillingProfileModel:
      type: object
      properties:
        total:
          type: integer
          description: Total number of billing profiles
        items:
          type: array
          items:
            $ref: '#/components/schemas/BillingProfileModel'
      description: >
        The total number of billing profiles available and a page of profiles
...
    JobModel:
      required:
        - id
        - job_status
        - status_code
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ShortIdProperty'
        description:
          type: string
          description: Description of the job's flight from description flight input
            parameter
        job_status:
          type: string
          description: Status of job
          enum:
            - running
            - succeeded
            - failed
        status_code:
          type: integer
          description: HTTP code
        submitted:
          type: string
          description: Timestamp when the flight was created
        completed:
          type: string
          description: Timestamp when the flight was completed; not present if not
            complete
        class_name:
          type: string
          description: Class name of the flight
      description: >
        Status of job
    ErrorModel:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        errorDetail:
          type: array
          items:
            type: string
      description: >
        Representation of an error that occurred while processing a request.
        Some errors provide error details. For example, validation of input may report several problems.
        Each of those are reported in a separate error detail.
...
    PolicyMemberRequest:
      required:
        - email
      type: object
      properties:
        email:
          type: string
      description: >
        email of user or group to add to policy
...
    PolicyModel:
      type: object
      properties:
        name:
          type: string
        members:
          type: array
          items:
            type: string
      description: >
        Describes a policy
...
    WorkspacePolicyModel:
      type: object
      properties:
        workspaceId:
          $ref: '#/components/schemas/UniqueIdProperty'
        workspaceName:
          type: string
        workspaceNamespace:
          type: string
        workspaceLink:
          type: string
        workspacePolicies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyModel'
    InaccessibleWorkspacePolicyModel:
      type: object
      properties:
        workspaceId:
          $ref: '#/components/schemas/UniqueIdProperty'
        workspacePolicies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyModel'
        error:
          $ref: "#/components/schemas/ErrorModel"
...
    PolicyResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyModel'
        authDomain:
          type: array
          items:
            description: The Sam user groups that define this auth domain
            type: string
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/WorkspacePolicyModel'
        inaccessibleWorkspaces:
          type: array
          items:
            $ref: '#/components/schemas/InaccessibleWorkspacePolicyModel'
      description: >
        email of user or group to add to policy
...
    CloudPlatform:
      type: string
      enum: [ gcp, azure ]
      default: gcp
      description: Cloud platforms supported by TDR.
...
  parameters:
    Id:
      name: id
      in: path
      description: A UUID to used to identify an object in the repository
      required: true
      schema:
        type: string
        format: uuid
...
    ProfilePolicyName:
      name: policyName
      in: path
      description: The relevant policy
      required: true
      schema:
        type: string
        enum:
          - owner
          - user
...
  securitySchemes:
    googleoauth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
            https://www.googleapis.com/auth/cloud-billing.readonly: cloud billing authorization
    oidc:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization